Fixes, refactor, fix search add single domain to cart
