+++ .env.example
@@ -7,6 +7,8 @@ NEXT_PUBLIC_BASE_URL="http://localhost:3443"
 EFP_API_URL=""
 NEXT_PUBLIC_EFP_API_URL=""
 
+NEXT_PUBLIC_ENS_SUBGRAPH_API_KEY=""
+
 NEXT_TELEMETRY_DISABLED="1"
 
 ANKR_ID=""
