+++ src/api/fetchProfile.ts
@@ -1,5 +1,5 @@
 import { isAddress, type Address } from 'viem'
-import type { ProfileResponse } from './requests'
+import type { ProfileResponse } from '#/types/requests'
 import { resolveENSAddress, resolveENSProfile } from '#/utils/resolveENS'
 
 const fetchUserProfile = async (addressOrName: string) => {
