+++ src/components/list-settings/hooks/use-save-list-settings.ts
@@ -5,8 +5,8 @@ import { isAddress, type Chain, type Address, encodePacked } from 'viem'
 
 import { useCart } from '#/contexts/cart-context'
 import { Step } from '#/components/checkout/types'
-import type { ProfileDetailsResponse } from '#/api/requests'
 import { useEFPProfile } from '#/contexts/efp-profile-context'
+import type { ProfileDetailsResponse } from '#/types/requests'
 import { efpListRecordsAbi, efpListRegistryAbi } from '#/lib/abi'
 import { generateListStorageLocationSlot } from '#/app/efp/utilities'
 import { coreEfpContracts, ListRecordContracts } from '#/lib/constants/contracts'
